backend/requirements.txt
argcomplete==2.0.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==22.2.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.10.3
blinker==1.5
Brlapi==0.8.4
Brotli==1.0.9
certifi==2022.9.24
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
cupshelpers==1.0
daphne==4.1.2
dbus-python==1.3.2
distro==1.8.0
Django==5.0.4
django-cors-headers==3.14.0
django-filter==23.1
django-guardian==2.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fuse-python==1.0.5
ghp-import==2.1.0
gpg==1.18.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
livereload==2.6.3
louis==3.24.0
lunr==0.6.2
Markdown==3.4.1
MarkupSafe==2.1.2
mergedeep==1.3.4
mkdocs==1.4.2
more-itertools==8.10.0
msgpack==1.0.8
mutagen==1.46.0
nltk==3.8
numpy==1.24.2
oauthlib==3.2.2
olefile==0.46
packaging==23.0
Pillow==9.4.0
pipx==1.1.0
pluggy==1.0.0+repack
psutil==5.9.4
psycopg2==2.9.5
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.20.1
pycparser==2.22
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.45.2
Pygments==2.14.0
PyGObject==3.42.2
pyinotify==0.9.6
PyJWT==2.6.0
pylibacl==0.7.0
pyOpenSSL==24.1.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-sip==12.11.1
PySimpleSOAP==1.16.2
pysmbc==1.0.23
pytest==7.2.1
python-apt==2.6.0
python-dateutil==2.8.2
python-debian==0.1.49
python-debianbts==4.0.1
pytz==2022.7.1
pyxattr==0.8.1
pyxdg==0.28
PyYAML==6.0
pyyaml_env_tag==0.1
redis==5.0.4
regex==2022.10.31
reportbug==12.0.0
requests==2.28.1
requests-oauthlib==2.0.0
service-identity==24.1.0
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.2
tornado==6.2
tqdm==4.64.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.12
userpath==1.8.0
wadllib==1.3.6
watchdog==2.2.1
websockets==10.4
xdg==5
yt-dlp==2023.3.4
zope.interface==6.3
